+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.31.0
   - feat: Add `NesBadge` widget that can display a badge indicator on any widget.
   - feat: Add `NesIcons.selection`.
+  - feat: Add `NesIcons.rotateFlipped`.
   - chore: Update `google_fonts` to latest
   - feat: Allow NesBottomSheet to have absolute height
   - feat: Add `NesIcons.cisel`.