+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - feat: Add `NesIcons.fillBucket`.
   - feat: Add `NesIcons.transform`.
   - feat: Add `NesIcons.shrink`.
+  - feat: Add `NesIcons.palette`.
 
 # 0.30.0
  - feat: Add `NesShakeEffect` widget with declarative `shaking` bool and optional `NesShakeEffectController`.