+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.32.0
+  - feat: Add `NesIcons.trophy`.
+
 # 0.31.0
   - feat: Add `NesBadge` widget that can display a badge indicator on any widget.
   - feat: Add `NesIcons.selection`.